fuse stats header hooks
